refactor(desktop): extract diff pane section tracking into a hook
Move the ~60 lines of active-section detection out of DiffPane.tsx into a
co-located useDiffActiveSection hook. DiffPane is a hot file; keeping the
feature's logic in its own module shrinks this branch's footprint there and
reduces future merge-conflict surface.
