+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/DiffPane/hooks/useDiffActiveSection/index.ts
@@ -1,0 +1,1 @@
+export { useDiffActiveSection } from "./useDiffActiveSection";
