+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/DiffPane/hooks/useDiffActiveSection/useDiffActiveSection.ts
@@ -1,0 +1,107 @@
+import type { CodeViewItem } from "@pierre/diffs";
+import type { CodeViewHandle } from "@pierre/diffs/react";
+import {
+	type RefObject,
+	useCallback,
+	useEffect,
+	useMemo,
+	useRef,
+	useState,
+} from "react";
+import type { ChangesetFile } from "../../../../../useChangeset";
+import type { DiffAnnotationMetadata } from "../useDiffAnnotations";
+
+type GroupKind = ChangesetFile["source"]["kind"];
+
+export interface DiffSection {
+	kind: GroupKind;
+	count: number;
+}
+
+interface UseDiffActiveSectionOptions {
+	codeViewRef: RefObject<CodeViewHandle<DiffAnnotationMetadata> | null>;
+	items: CodeViewItem<DiffAnnotationMetadata>[];
+	fileByItemId: ReadonlyMap<string, ChangesetFile>;
+	files: ChangesetFile[];
+}
+
+interface UseDiffActiveSectionResult {
+	/** Source group of the topmost visible file, for the pinned section bar. */
+	currentSection: DiffSection | null;
+	/** Wire to CodeView's `onScroll`. */
+	onScroll: () => void;
+}
+
+/**
+ * Tracks which source group (unstaged / staged / committed …) the topmost
+ * visible file belongs to, so the diff pane can pin a single section bar and
+ * update it as the user scrolls across group boundaries.
+ */
+export function useDiffActiveSection({
+	codeViewRef,
+	items,
+	fileByItemId,
+	files,
+}: UseDiffActiveSectionOptions): UseDiffActiveSectionResult {
+	const { sectionByItemId, firstSection } = useMemo(() => {
+		const counts = new Map<GroupKind, number>();
+		for (const file of files) {
+			counts.set(file.source.kind, (counts.get(file.source.kind) ?? 0) + 1);
+		}
+		const byItemId = new Map<string, DiffSection>();
+		let first: DiffSection | null = null;
+		for (const item of items) {
+			const kind = fileByItemId.get(item.id)?.source.kind;
+			if (!kind) continue;
+			const entry: DiffSection = { kind, count: counts.get(kind) ?? 0 };
+			byItemId.set(item.id, entry);
+			first ??= entry;
+		}
+		return { sectionByItemId: byItemId, firstSection: first };
+	}, [items, fileByItemId, files]);
+
+	const [topItemId, setTopItemId] = useState<string | null>(null);
+	const currentSection =
+		(topItemId != null ? sectionByItemId.get(topItemId) : undefined) ??
+		firstSection;
+
+	// Track the topmost item by geometry rather than `getRenderedItems()[0]`,
+	// which can lag by the virtualization buffer. Items stack top→bottom with
+	// monotonically increasing offsets, so binary-search for the last one whose
+	// top has passed the viewport top: its header is the pinned one.
+	const rafRef = useRef<number | null>(null);
+	const updateActiveSection = useCallback(() => {
+		rafRef.current = null;
+		const instance = codeViewRef.current?.getInstance();
+		if (!instance) return;
+		const scrollTop = instance.getScrollTop();
+		let lo = 0;
+		let hi = items.length - 1;
+		let nextTopId: string | null = null;
+		while (lo <= hi) {
+			const mid = (lo + hi) >> 1;
+			const top = instance.getTopForItem(items[mid].id);
+			if (top != null && top <= scrollTop + 1) {
+				nextTopId = items[mid].id;
+				lo = mid + 1;
+			} else {
+				hi = mid - 1;
+			}
+		}
+		setTopItemId((prev) => (prev === nextTopId ? prev : nextTopId));
+	}, [codeViewRef, items]);
+
+	// Coalesce bursts of scroll events into one measurement per frame.
+	const onScroll = useCallback(() => {
+		if (rafRef.current != null) return;
+		rafRef.current = requestAnimationFrame(updateActiveSection);
+	}, [updateActiveSection]);
+
+	useEffect(() => {
+		return () => {
+			if (rafRef.current != null) cancelAnimationFrame(rafRef.current);
+		};
+	}, []);
+
+	return { currentSection, onScroll };
+}
